VeriScan_Project/requirements.txt
streamlit
torch
transformers
Pillow
numpy
scikit-image